Fix overly simplistic Rotor mapping
Rotor mapping was not considering in the "outbound" current flow during
encoding creating a bi-directional mapping of values. Instead it was
using the same "wiring" on both inbound and outbound directions.

Rotor related structs have been updated to include a "in" and "out"
mapping to accommodate this. The mapping API has changes to include a
post-fixed indicator of which direction the mapping is going. E.g.,
`map_in` and `map_out`. The "in" direction is from the keyboard to the
reflector and "out" is from the reflector to the light board.

Rotor mapping now properly considers the rotors location. Previously the
mapping was only handling the "wiring" of the rotor tyre. The current
location of the rotor must be considered for a "proper" translation.
E.g., an 'E' output when the rotor is in position 1 is actually an
output at position 'D'.

Nearly all the rotor related tests have been updated due to the above
fixes. A new test has been added to ensure that all the tyre alphabet
mappings are properly formatted. I.e., a given IN and OUT alphabet
together form a bi-directional mapping.