Fix incorrect rotor ring handling
Update Rotor handling of ring settings to address incorrect translation
issues. Rotor ring settings cause an adjustment between a rotor's
"external" input values and the "internal" wiring that they map to.
This effectively just changes where in the mapping tables we should look
for values. However, it's important that we consider the ring offset when
calculating the proper output from a rotor.

The previous implementation, in addition to being overly simplified due
to misunderstanding examples listed in various documentation sources,
was also offsetting the ring value in reverse compared to the "common"
parlance. A ring value of 1 is now equivalent to a right rotation of the
mapping array. E.g., Z's mapping becomes A's mapping with a ring loc
value of 1. Note that this can make the mapping math look backwards.

Added and updated some tests as necessary. Note that some of our old
tests were just wrong or not particularly useful given the changes made.
Should look into adding additional tests when a future cleanup pass is
done. Seriously, the amount of type casts is embarrassing at the
moment.