Add initial CLI
Add initial (messy) CLI to support translating text via the `enigma`
binary. The CLI expects to receive a string defining the enigma machine
configuration along with input via stdin or a flag.

This iteration includes a fair bit of logic for checking the validity of
the passed configuration. Need to consider relocating that to the
appropriate modules and adding tests. Additionally, should consider
pulling out the argument and config parsing code into separate modules.